+++ Bootstrap.php
@@ -29,6 +29,10 @@ class Shopware_Plugins_Frontend_jtlconnector_Bootstrap extends Shopware_Componen
         $this->runAutoload();
 
         parent::__construct($name, $info);
+
+        /** @var Enlight_Event_EventManager $eventManager */
+        $eventManager = Shopware()->Container()->get('events');
+        $eventManager->addSubscriber(new \jtl\Connector\Shopware\Subscriber\Translation());
     }
     
     public function getCapabilities()
@@ -296,6 +300,8 @@ class Shopware_Plugins_Frontend_jtlconnector_Bootstrap extends Shopware_Componen
             case '2.2.1.3':
             case '2.2.1.4':
             case '2.2.2':
+            case '2.2.3':
+            case '2.2.3.1':
                 Shopware()->Db()->query("DROP TABLE IF EXISTS `jtl_connector_category_level`");
                 break;
             default:
@@ -375,7 +381,7 @@ class Shopware_Plugins_Frontend_jtlconnector_Bootstrap extends Shopware_Componen
     {
         return dirname(__FILE__) . '/Connector.php';
     }
-    
+
     private function runAutoload()
     {
         // Tmp directory fallback